1.2 Release
This is the 1.2 beta release, with the following changes:
-Removed raworiginal.s for copyright reasons
-Added internalmem.dat so builds will work fine
-Added autowrite/fix to file

TODO:
-Despite autowrite has been implemented, tool is not reloading file,
therefore you will have to fix slot1/slot2 checksums, close the tool,
reopen it and fix file checksum. If you don't close the tool before
patching file checksum it will not be correct, therefore save will be
marked as corrupt.